Implemented Strategy Pattern
