+++ src/main/java/ca/concordia/pivottable/servicelayer/impl/UserManagementServiceImpl.java
@@ -27,7 +27,8 @@ public class UserManagementServiceImpl implements UserManagementService
 	/**
 	 * User database object used for performing data processing operations.
 	 */
-	private UserDataAccess userDatabase = new UserDataAccessImpl();
+	//TODO need correct implementation here
+	private UserDataAccess userDatabase = new UserDataAccessImpl(null, null);
 	
 	/**
 	 * Adds a new user's credentials (username in lower case and password hash) to the users database.
